+++ src/main/java/nextstep/refactoring/laddergame/engine/player/PlayersDto.java
@@ -1,9 +1,17 @@
 package nextstep.refactoring.laddergame.engine.player;
 
+import java.util.Collections;
 import java.util.List;
 
-public interface PlayersDto {
+public class PlayersDto {
 
-    List<String> nameList();
+    private final List<String> nameList;
 
+    public PlayersDto(List<String> nameList) {
+        this.nameList = nameList;
+    }
+
+    public List<String> nameList() {
+        return Collections.unmodifiableList(nameList);
+    }
 }
