+++ src/main/java/nextstep/refactoring/laddergame/engine/reward/RewardsDto.java
@@ -1,9 +1,18 @@
 package nextstep.refactoring.laddergame.engine.reward;
 
+import java.util.Collections;
 import java.util.List;
 
-public interface RewardsDto {
+public class RewardsDto {
 
-    List<String> rewards();
+    private final List<String> rewardList;
+
+    public RewardsDto(List<String> rewardList) {
+        this.rewardList = rewardList;
+    }
+
+    public List<String> rewards() {
+        return Collections.unmodifiableList(rewardList);
+    }
 
 }
